ref: move utils to worker
