+++ src/index.ts
@@ -1,4 +1,3 @@
 export * from './contexts';
 export * from './hooks';
-export * from './utils';
 export * from './consumer';
